新增lua memory snapshoot

+++ src/cJSON.c
@@ -141,7 +141,7 @@ static int update(printbuffer *p)
 	char *str;
 	if (!p || !p->buffer) return 0;
 	str=p->buffer+p->offset;
-	return p->offset+strlen(str);
+	return p->offset+(int)strlen(str);
 }
 
 #ifdef _MSC_VER
@@ -546,7 +546,7 @@ static char *print_array(cJSON *item,int depth,int fmt,printbuffer *p)
 		{
 			ret=print_value(child,depth+1,fmt,0);
 			entries[i++]=ret;
-			if (ret) len+=strlen(ret)+2+(fmt?1:0); else fail=1;
+			if (ret) len+=(int)strlen(ret)+2+(fmt?1:0); else fail=1;
 			child=child->next;
 		}
 		
@@ -731,7 +731,7 @@ static char *print_object(cJSON *item,int depth,int fmt,printbuffer *p)
 		{
 			names[i]=str=print_string_ptr(child->string,0);
 			entries[i++]=ret=print_value(child,depth,fmt,0);
-			if (str && ret) len+=strlen(ret)+strlen(str)+2+(fmt?2+depth:0); else fail=1;
+			if (str && ret) len+=(int)strlen(ret)+(int)strlen(str)+2+(fmt?2+depth:0); else fail=1;
 			child=child->next;
 		}
